fix: category service NoneType client + match_score float→int (closes #34, closes #35)
Bug 1 (#34): category_service.py created its own Supabase admin client at
module level, falling back to None if SUPABASE_SERVICE_ROLE_KEY wasn't set.
This broke the AI state/local grant scraper (0 categories found) and the
/api/categories endpoint. Fix: get_category_service() now accepts an optional
client parameter, and startup_event() initializes it with the admin client.

Bug 2 (#35): calculate_match_score() returned floats due to context_matches * 1.5
but the DB column is integer. Grants with float scores failed to insert:
'invalid input syntax for type integer: "32.0"'. Fix: wrap returns in int().

+++ backend/match_scoring.py
@@ -93,9 +93,9 @@ def calculate_match_score(grant: Dict[str, Any], rfp_similarities: List[Dict[str
     # Calculate keyword score (max 30 points)
     # Require at least 2 core keywords for decent score
     if core_matches >= 2:
-        keyword_score = min(30, (core_matches * 6) + (context_matches * 1.5))
+        keyword_score = int(min(30, (core_matches * 6) + (context_matches * 1.5)))
     elif core_matches == 1:
-        keyword_score = min(12, core_matches * 6)
+        keyword_score = int(min(12, core_matches * 6))
     else:
         keyword_score = 0  # No core keywords = very low relevance
 
@@ -168,7 +168,7 @@ def calculate_match_score(grant: Dict[str, Any], rfp_similarities: List[Dict[str
     
     # Add 10 point boost to all scores
     final_score = final_score + 10
-    return min(100, max(0, final_score))
+    return int(min(100, max(0, final_score)))
 
 
 def get_score_breakdown(grant: Dict[str, Any], rfp_similarities: List[Dict[str, Any]] = []) -> Dict[str, Any]:
